Pull out client fixture
