+++ tests/conftest.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+import pytest
+from fixtures.client import channel_maker, client, patch_discord, session  # type: ignore
