+++ tests/test_spellbot.py
@@ -6,14 +6,13 @@ from datetime import datetime, timedelta
 from os import chdir
 from pathlib import Path
 from subprocess import run
-from unittest.mock import MagicMock
 from warnings import warn
 
 import pytest
 import pytz
 import toml
+from fixtures.client import simulate_user_leaving_server  # type: ignore
 from helpers.constants import (  # type:ignore
-    CLIENT_AUTH,
     CLIENT_TOKEN,
     FIXTURES_ROOT,
     REPO_ROOT,
@@ -21,19 +20,15 @@ from helpers.constants import (  # type:ignore
     SRC_DIRS,
 )
 from mocks.discord import (  # type: ignore
-    AsyncMock,
     MockAttachment,
     MockChannel,
-    MockDM,
     MockMessage,
     MockPayload,
-    MockTextChannel,
 )
 from mocks.users import (  # type: ignore
     ADAM,
     ADMIN,
     ADMIN_ROLE,
-    ALL_USERS,
     AMY,
     BOT,
     BUDDY,
@@ -51,30 +46,6 @@ import spellbot
 from spellbot.assets import load_strings
 from spellbot.constants import THUMB_URL
 from spellbot.data import Event, Game, User
-
-
-class MockDiscordClient:
-    def __init__(self, **kwargs):
-        self.user = ADMIN
-        self.channels = []
-
-    def get_user(self, user_id):
-        for user in ALL_USERS:
-            if user.id == user_id:
-                return user
-
-    async def fetch_user(self, user_id):
-        return self.get_user(user_id)
-
-    def get_channel(self, channel_id):
-        for channel in self.channels:
-            if channel.id == channel_id:
-                return channel
-        return None
-
-    async def fetch_channel(self, channel_id):
-        return self.get_channel(channel_id)
-
 
 ##############################
 # Test Suite Utilities
@@ -103,6 +74,7 @@ def is_discord_file(obj):
     return (obj.__class__.__name__) == "File"
 
 
+@pytest.mark.usefixtures("client")
 def user_has_game(client, user):
     session = client.data.Session()
     db_user = session.query(User).filter(User.xid == user.id).first()
@@ -111,6 +83,7 @@ def user_has_game(client, user):
     return rvalue
 
 
+@pytest.mark.usefixtures("client")
 def game_json_for(client, user):
     session = client.data.Session()
     db_user = session.query(User).filter(User.xid == user.id).first()
@@ -119,6 +92,7 @@ def game_json_for(client, user):
     return rvalue
 
 
+@pytest.mark.usefixtures("client")
 def game_embed_for(client, user, ready, message=None):
     session = client.data.Session()
     db_user = session.query(User).filter(User.xid == user.id).first()
@@ -131,6 +105,7 @@ def game_embed_for(client, user, ready, message=None):
     return rvalue.to_dict() if rvalue else None
 
 
+@pytest.mark.usefixtures("client")
 def all_games(client):
     session = client.data.Session()
     games = session.query(Game).all()
@@ -139,6 +114,7 @@ def all_games(client):
     return rvalue
 
 
+@pytest.mark.usefixtures("client")
 def all_events(client):
     session = client.data.Session()
     events = session.query(Event).all()
@@ -152,55 +128,9 @@ def all_events(client):
 ##############################
 
 
-@pytest.fixture
-def patch_discord():
-    orig = spellbot.SpellBot.__bases__
-    spellbot.SpellBot.__bases__ = (MockDiscordClient,)
-    yield
-    spellbot.SpellBot.__bases__ = orig
-
-
 @pytest.fixture(autouse=True, scope="session")
 def set_random_seed():
     random.seed(0)
-
-
-@pytest.fixture
-def client(monkeypatch, mocker, patch_discord, tmp_path):
-    # Define which users are on this Discord server.
-    global ALL_USERS
-    ALL_USERS = SERVER_MEMBERS + [PUNK, BOT]
-
-    # Keep track of strings used in the test suite.
-    monkeypatch.setattr(spellbot, "s", S_SPY)
-
-    # Don't actually begin background tasks during tests.
-    monkeypatch.setattr(spellbot.SpellBot, "_begin_background_tasks", MagicMock())
-
-    # Fallback to using sqlite for tests, but use the environment variable if it's set.
-    (tmp_path / "spellbot.db").touch()
-    connection_string = f"sqlite:///{tmp_path}/spellbot.db"
-    db_url = spellbot.get_db_url("TEST_SPELLBOT_DB_URL", connection_string)
-    bot = spellbot.SpellBot(
-        token=CLIENT_TOKEN, auth=CLIENT_AUTH, db_url=db_url, mock_games=True
-    )
-
-    # Each test should have a clean slate. If we're using sqlite this is ensured
-    # automatically as each test will create its own new spellbot.db file. With other
-    # databases we'll have to manually clean out any existing data before each
-    # test as the previous tests could have left data behind.
-    for table in bot.data.metadata.tables.keys():
-        bot.data.conn.execute(f"DELETE FROM {table};")
-
-    # Make sure that all users have their send calls reset between client tests.
-    for user in ALL_USERS:
-        user.sent = AsyncMock()
-
-    yield bot
-
-    # For sqlite closing the connection when we're done isn't necessary, but for other
-    # databases our test suite can quickly exhaust their connection pools.
-    bot.data.conn.close()
 
 
 SNAPSHOTS_USED = set()
@@ -221,47 +151,14 @@ def snap(snapshot):
     return match
 
 
-@pytest.fixture
-def session(client):
-    """
-    The client creates a session when processing a command,
-    we have to create one ourselves when not in a command context.
-    """
-    return client.data.Session()
-
-
-@pytest.fixture
-def channel_maker(client):
-    """Use this fixture to create channels so that we can hook them up to the client."""
-
-    class MockChannelFactory:
-        def __init__(self, next_channel_id):
-            self.next_channel_id = next_channel_id
-
-        def text(self, name="a-text-channel", members=SERVER_MEMBERS):
-            channel = MockTextChannel(self.next_channel_id, name, members)
-            client.channels.append(channel)  # not something you can do with a real client
-            self.next_channel_id += 1
-            return channel
-
-        def dm(self):
-            channel = MockDM(self.next_channel_id)
-            self.next_channel_id += 1
-            return channel
-
-        def make(self, channel_type):
-            if channel_type == "dm":
-                return self.dm()
-            return self.text()
-
-    return MockChannelFactory(6500)
-
-
 ##############################
 # Test Suites
 ##############################
 
 
+@pytest.mark.usefixtures("client")
+@pytest.mark.usefixtures("channel_maker")
+@pytest.mark.usefixtures("session")
 @pytest.mark.asyncio
 class TestSpellBot:
     async def test_init(self, client, channel_maker):
@@ -1273,9 +1170,7 @@ class TestSpellBot:
             f" next run `!begin {event_id}` to start the event."
         )
 
-        # Simulate AMY leaving this Discord server
-        global ALL_USERS
-        ALL_USERS = [user for user in ALL_USERS if user != AMY]
+        simulate_user_leaving_server(AMY)
 
         await client.on_message(MockMessage(author, channel, f"!begin {event_id}"))
         assert channel.last_sent_response == (
@@ -1778,9 +1673,7 @@ class TestSpellBot:
         await client.on_message(MockMessage(GUY, channel, "!lfg size:2"))
         message = channel.last_sent_message
 
-        # Simulate GUY leaving this Discord server
-        global ALL_USERS
-        ALL_USERS = [user for user in ALL_USERS if user != GUY]
+        simulate_user_leaving_server(GUY)
 
         payload = MockPayload(
             user_id=ADAM.id,
@@ -1979,9 +1872,7 @@ class TestSpellBot:
 
         assert user_has_game(client, GUY)
 
-        # Simulate GUY leaving this Discord server
-        global ALL_USERS
-        ALL_USERS = [user for user in ALL_USERS if user != GUY]
+        simulate_user_leaving_server(GUY)
 
         freezer.move_to(NOW + timedelta(days=3))
         await client.cleanup_expired_games()
@@ -2261,6 +2152,8 @@ class TestMigrations:
         reverse_all(connection, connection_url)
 
 
+@pytest.mark.usefixtures("client")
+@pytest.mark.usefixtures("channel_maker")
 class TestCodebase:
     def test_mypy(self):
         """Checks that the Python codebase passes mypy static analysis checks."""
